# src/main/resources/static/openapi.yml
openapi: 3.0.1
info:
  title: SO (Stack Overflow clone) API
  description: This is a clone of Stack Overflow, providing a platform for users to ask and answer technical questions.
  version: 1.0.0
paths:
  /api/v1/auth/register:
    post:
      summary: Register a new user
      description: This endpoint allows you to register a new user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
        '400':
          description: Bad request
  /api/v1/auth/login:
    post:
      summary: Log user in
      description: This endpoint allows a user to login.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User logged in successfully
        '401':
          description: Unauthorized
  /api/v1/questions:
    post:
      summary: Ask a question
      description: This endpoint allows a user to ask a question.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Question'
      responses:
        '201':
          description: Question asked successfully
        '400':
          description: Bad request
  /api/v1/questions/{questionId}:
    put:
      summary: Update a question
      description: This endpoint allows a user to update a question.
      parameters:
          - name: questionId
            in: path
            required: true
            description: ID of the question to update
            schema:
               type: string
               format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Question'
      responses:
        '200':
          description: Question updated successfully
        '400':
          description: Bad request
        '404':
          description: Question not found

components:
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    Question:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        userId:
          type: string
          format: uuid
      required:
        - title
        - content
        - userId